requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.2
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-select2==7.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
soupsieve==2.4
sqlparse==0.4.3
channels==3.0.5
channels-redis==3.4.1
djangochannelsrestframework==1.1.0 
daphne==3.0.2
aioredis==1.3.1
redis==4.5.4